+++ flink-core/src/main/java/org/apache/flink/configuration/JobManagerOptions.java
@@ -580,6 +580,12 @@ public class JobManagerOptions {
                                             code(SchedulerExecutionMode.REACTIVE.name()))
                                     .build());
 
+    /** Enum for controlling Task deployment mode. */
+    public enum TaskDeploymentMode {
+        SINGLE,
+        BATCH
+    }
+
     @Documentation.Section({Documentation.Sections.ALL_JOB_MANAGER})
     public static final ConfigOption<TaskDeploymentMode> TASK_DEPLOYMENT_MODE =
             key("jobmanager.execution.deploy-mode")
