--- flink-runtime/src/main/java/org/apache/flink/runtime/scheduler/BatchExecutionDeployManager.java
@@ -1,69 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.apache.flink.runtime.scheduler;
-
-import org.apache.flink.runtime.deployment.TaskDeploymentDescriptor;
-import org.apache.flink.runtime.executiongraph.ExecutionAttemptID;
-import org.apache.flink.runtime.jobmaster.LogicalSlot;
-import org.apache.flink.runtime.messages.Acknowledge;
-import org.apache.flink.runtime.taskmanager.TaskManagerLocation;
-
-import org.apache.commons.collections.MapUtils;
-
-import java.util.Map;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ConcurrentHashMap;
-
-/** Manager to manage deployment of batch executions. */
-public class BatchExecutionDeployManager {
-    private final Map<TaskManagerLocation, TaskBatchDeploymentGroup> taskManagerLocationMap;
-
-    private final Map<ExecutionAttemptID, CompletableFuture<Acknowledge>> resultFutureMap;
-
-    public BatchExecutionDeployManager() {
-        taskManagerLocationMap = new ConcurrentHashMap<>();
-        resultFutureMap = new ConcurrentHashMap<>();
-    }
-
-    public void registerTaskmanagerLocation(
-            LogicalSlot logicalSlot, TaskDeploymentDescriptor taskDeploymentDescriptor) {
-        final TaskManagerLocation taskManagerLocation = logicalSlot.getTaskManagerLocation();
-        taskManagerLocationMap
-                .computeIfAbsent(
-                        taskManagerLocation,
-                        t -> new TaskBatchDeploymentGroup(logicalSlot.getTaskManagerGateway()))
-                .addTaskDeploymentDescriptor(taskDeploymentDescriptor);
-        resultFutureMap.put(
-                taskDeploymentDescriptor.getExecutionAttemptId(), new CompletableFuture<>());
-    }
-
-    public Map<TaskManagerLocation, TaskBatchDeploymentGroup> getTaskManagerLocationMap() {
-        return taskManagerLocationMap;
-    }
-
-    public CompletableFuture<Acknowledge> getDeployResultFuture(
-            ExecutionAttemptID executionAttemptID) {
-        return resultFutureMap.get(executionAttemptID);
-    }
-
-    public Map<ExecutionAttemptID, CompletableFuture<Acknowledge>> getAllDeployResultFutures() {
-        return MapUtils.unmodifiableMap(resultFutureMap);
-    }
-}
