--- flink-runtime/src/main/java/org/apache/flink/runtime/scheduler/ExecutionDeployerFactoryLoader.java
@@ -1,42 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.apache.flink.runtime.scheduler;
-
-import org.apache.flink.configuration.Configuration;
-import org.apache.flink.configuration.JobManagerOptions;
-import org.apache.flink.configuration.TaskDeploymentMode;
-
-/** A utility class to load {@link ExecutionDeployer.Factory} from the configuration. */
-public class ExecutionDeployerFactoryLoader {
-
-    public static ExecutionDeployExecutor.Factory createExecutionDeployerFactory(
-            Configuration configuration) {
-        final TaskDeploymentMode taskDeploymentMode =
-                configuration.get(JobManagerOptions.TASK_DEPLOYMENT_MODE);
-        switch (taskDeploymentMode) {
-            case BATCH:
-                return new BatchExecutionDeployExecutor.Factory();
-            case SINGLE:
-                return new DefaultExecutionDeployExecutor.Factory();
-            default:
-                throw new UnsupportedOperationException("Not support taskDeploymentMode");
-        }
-    }
-}
